Add Stripe refund capabilities
We can refund payments directly via Stripe.  Make it so, and provide
error handling if the charge can't be refunded in that manner.  Note
that this requires the payment_extra attribute to contain the Stripe
charge ID; if it contains something else, it will fail! (payment_extra
can still be used for non-credit transactions.)